% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbma.R
\name{evalue_mbma}
\alias{evalue_mbma}
\title{E-value for meta-analysis with multiple biases}
\usage{
evalue_mbma(
  yi,
  vi,
  sei,
  selection_ratio,
  q = 0,
  cluster = 1:length(yi),
  biased = TRUE,
  favor_positive = TRUE,
  alpha_select = 0.05,
  ci_level = 0.95,
  small = TRUE,
  bias_grid_hi = 20,
  evalue_transformation = function(b) b + sqrt(b^2 - b)
)
}
\arguments{
\item{yi}{A vector of point estimates to be meta-analyzed.}

\item{vi}{A vector of estimated variances (i.e., squared standard errors) for
the point estimates.}

\item{sei}{A vector of estimated standard errors for the point estimates.
(Only one of \code{vi} or \code{sei} needs to be specified).}

\item{selection_ratio}{Ratio by which publication bias favors affirmative
studies (i.e., studies with p-values less than \code{alpha_select} and estimates
in the direction indicated by \code{favor_positive}).}

\item{q}{The attenuated value to which to shift the point estimate or CI.
Should be specified on the same scale as \code{dat$yi} (e.g., if
\code{dat$yi} is on the log-RR scale, then \code{q} should be as well).}

\item{cluster}{Vector of the same length as the number of rows in the data,
indicating which cluster each study should be considered part of (defaults
to treating studies as independent; i.e., each study is in its own cluster).}

\item{biased}{Boolean indicating whether each study is considered internally
biased; either single value used for all studies or a vector the same
length as the number of rows in the data (defaults to all studies).}

\item{favor_positive}{\code{TRUE} if publication bias are
assumed to favor significant positive estimates; \code{FALSE} if assumed to
favor significant negative estimates.}

\item{alpha_select}{Alpha level at which an estimate's probability of being
favored by publication bias is assumed to change (i.e.,
the threshold at which study investigators, journal editors, etc., consider
an estimate to be significant).}

\item{ci_level}{Confidence interval level (as proportion) for the corrected
point estimate. (The alpha level for inference on the corrected point
estimate will be calculated from \code{ci_level}.)}

\item{small}{Should inference allow for a small meta-analysis? We recommend
always using \code{TRUE}.}

\item{bias_grid_hi}{The largest value of \code{bias}, on the additive scale, that should be included
in the grid search. The bias has the same units as \code{yi}.}

\item{evalue_transformation}{TODO MM: Please see multi_evalue_example.R}
}
\description{
E-value for meta-analysis with multiple biases
}
\examples{
evalue_mbma(yi = meta_meat$yi,
            vi = meta_meat$vi,
            selection_ratio = 4,
            biased = !meta_meat$randomized)
}
