% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbma-package.R
\docType{package}
\name{mbma-package}
\alias{mbma}
\alias{mbma-package}
\title{mbma: Sensitivity analysis for the interactive effects of internal bias and publication bias in meta-analyses}
\description{
Meta-analyses can be compromised by studies' internal biases (e.g., confounding in nonrandomized studies) as well as publication bias. Per Mathur 2022 (\url{https://osf.io/u7vcb/}), this package conducts sensitivity analyses that address two questions: (1) "For a given severity of internal bias across studies and of publication bias, how much could the results change?"; and (2) "For a given severity of publication bias, how severe would internal bias have to be, hypothetically, to attenuate the results to the null or by a given amount?"
}
\author{
\strong{Maintainer}: Mika Braginsky \email{mika.br@gmail.com}

Authors:
\itemize{
  \item Maya Mathur \email{mmathur@stanford.edu}
}

}
\keyword{internal}
