% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta_meat}
\alias{meta_meat}
\title{Meta-analysis about meat consumption}
\format{
A data frame with 100 rows and 4 columns:
\describe{
  \item{yi}{Point estimate on log-risk ratio scale}
  \item{vi}{Variance of point estimate}
  \item{cluster}{Paper that contributed the point estimate}
  \item{randomized}{Logical indicating whether study was randomized}
}
}
\usage{
meta_meat
}
\description{
Meta-analysis of the effectiveness of educational behavior interventions that
attempt to reduce meat consumption by appealing to animal welfare.
}
\references{
\insertRef{mathur2021}{mbma}
}
\keyword{datasets}
