% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalue.R
\name{multibias_evalue}
\alias{multibias_evalue}
\title{E-value for meta-analysis with multiple biases}
\usage{
multibias_evalue(
  yi,
  vi,
  sei,
  cluster = 1:length(yi),
  biased = TRUE,
  selection_ratio,
  q = 0,
  favor_positive = TRUE,
  alpha_select = 0.05,
  ci_level = 0.95,
  small = TRUE,
  bias_max = 20,
  assumed_bias_type = NULL
)
}
\arguments{
\item{yi}{A vector of point estimates to be meta-analyzed.}

\item{vi}{A vector of estimated variances (i.e., squared standard errors) for
the point estimates.}

\item{sei}{A vector of estimated standard errors for the point estimates.
(Only one of `vi` or `sei` needs to be specified).}

\item{cluster}{Vector of the same length as the number of rows in the data,
indicating which cluster each study should be considered part of (defaults
to treating studies as independent; i.e., each study is in its own cluster).}

\item{biased}{Boolean indicating whether each study is considered internally
biased; either single value used for all studies or a vector the same
length as the number of rows in the data (defaults to all studies).}

\item{selection_ratio}{Ratio by which publication bias favors affirmative
studies (i.e., studies with p-values less than `alpha_select` and
estimates in the direction indicated by `favor_positive`).}

\item{q}{The attenuated value to which to shift the point estimate or CI.
Should be specified on the same scale as `yi` (e.g., if
`yi` is on the log-RR scale, then `q` should be as well).}

\item{favor_positive}{`TRUE` if publication bias are
assumed to favor significant positive estimates; `FALSE` if assumed to
favor significant negative estimates.}

\item{alpha_select}{Alpha level at which an estimate's probability of being
favored by publication bias is assumed to change (i.e.,
the threshold at which study investigators, journal editors, etc., consider
an estimate to be significant).}

\item{ci_level}{Confidence interval level (as proportion) for the corrected
point estimate. (The alpha level for inference on the corrected point
estimate will be calculated from `ci_level`.)}

\item{small}{Should inference allow for a small meta-analysis? We recommend
always using `TRUE`.}

\item{bias_max}{The largest value of \code{bias}, on the additive scale,
that should be included in the grid search. The bias has the same units as
\code{yi}.}

\item{assumed_bias_type}{List of biases to consider for computing evalues
(objects of \code{bias} as returned by \code{EValue::confounding()},
\code{EValue::selection()}, \code{EValue::misclassification()}) (defaults
to NULL, i.e. agnostic as to the nature of the internal bias). If not NULL,
the \code{yi} argument must be on the log-RR scale (if \code{yi} is not
already on that scale, use \code{EValue::convert_measures()} to make it
so).}
}
\description{
E-value for meta-analysis with multiple biases
}
\details{
For more on the functions passed as \code{assumed_bias_type}, see the
EValue package vignette on multiple biases:
\url{https://cran.r-project.org/web/packages/EValue/vignettes/multiple-bias.html}.
}
\examples{
multibias_evalue(yi = meta_meat$yi,
                 vi = meta_meat$vi,
                 selection_ratio = 4,
                 biased = !meta_meat$randomized)

# specify confounding as internal bias
multibias_evalue(yi = meta_meat$yi,
                 vi = meta_meat$vi,
                 selection_ratio = 4,
                 biased = !meta_meat$randomized,
                 assumed_bias_type = list(EValue::confounding()))
}
\references{
\insertRef{mathur2022}{multibiasmeta}

\insertRef{ding2016}{multibiasmeta}

\insertRef{smith2019}{multibiasmeta}

\insertRef{vanderweele2019}{multibiasmeta}
}
