% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multibias.R
\name{robu_ci}
\alias{robu_ci}
\title{Add CI with specified ci_level to reg_table output from robumeta::robu()}
\usage{
robu_ci(reg_table, ci_level)
}
\description{
Add CI with specified ci_level to reg_table output from robumeta::robu()
}
\keyword{internal}
