% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbma-package.R
\docType{package}
\name{mbma-package}
\alias{mbma}
\alias{mbma-package}
\title{mbma: TODO}
\description{
TODO. TODO. TODO.
}
\author{
\strong{Maintainer}: Mika Braginsky \email{mika.br@gmail.com}

Authors:
\itemize{
  \item Maya Mathur \email{mmathur@stanford.edu}
}

}
\keyword{internal}
