% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbma.R
\name{robu_ci}
\alias{robu_ci}
\title{Add custom CI to robumeta::robu() reg_table output}
\usage{
robu_ci(reg_table, ci_level)
}
\description{
Add custom CI to robumeta::robu() reg_table output
}
\keyword{internal}
